requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==41.0.7

# LangGraph and LangChain
langgraph==0.0.55
langchain==0.1.6
langchain-openai==0.0.5
langchain-core==0.1.17

# Utilities
python-dotenv==1.0.0
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
